patternDecoder:
General: Improved long message handling. If message buffer is full, unneeded Data is truncated.

For Manchester this means, if last entry in message buffer is a valid manchester pulse, we will truncate all data from start to the first valid manchester. This helps to capture long signals which are not detected from the beginning.

For signals with sync, if message end (start from a repeated message) is not found, but a message start, we will truncate everything bevore start in message buffer if the message buffer is full and try to gather some more data.

Reset isn't anymore done in doDetectwoSync function.

ManchesterpatternDecoder::doDecode() : fixed a bug,detecting the start of manchester correctly