patternDecoder.cpp
Added a compress_pattern Function. This is called just bevore we print out our message.
If there are equal messages over moving average, they will be combined to one pattern.

preamble and postambe string are now vars in patternDecoder. This prevents some errors within the memory manager. The space for this strings (90 byte) is not freed and allocated as bevore.

Added a class patternO for future usage and replacement of 2d pattern array