SignalDetectorClass : Enhance Manchester detection
Resetting mc signal vars, every time the class is resetted instead of doing this when it is initalized only.
Corrected message truncate for mc messages. Do it only, if the buffer is full and we have manchester data, not when the buffer is not full.
Reset manchester decoder, when adding new pattern / overwriting old ones.