signalDecoder.cpp
When adding a pattern check if we overwrite more than 2 entries and call processMessage if so. This helps to provide messages, which may fit excactly in the pattern buffer.

Fixed resetting values in ManchesterpatternDecoder

Fixed a bug  overflow a int8 var in isManchester calculating the equal_cnt. If first value was negative, then the var was overflowed.